+++ apps/client/src/views/home/ui/HomePage/index.tsx
@@ -2,7 +2,7 @@ import { UserRoleType } from '@repo/shared/types';
 import { cn } from '@repo/shared/utils';
 
 import { getApiKey, getAvailableScope } from '@/views/home';
-import { ApiKeyCard, ApiKeyHeader } from '@/widgets/home';
+import { ApiKeyDisplay, ApiKeyForm, ApiKeyHeader } from '@/widgets/home';
 
 const HomePage = async () => {
   const userRole: UserRoleType = 'USER';
@@ -17,11 +17,14 @@ const HomePage = async () => {
       <main className={cn('container mx-auto px-4 py-16')}>
         <div className={cn('mx-auto max-w-2xl')}>
           <ApiKeyHeader />
-          <ApiKeyCard
-            initialApiKeyData={initialApiKeyData}
-            initialAvailableScope={initialAvailableScope}
-            userRole={userRole}
-          />
+          <div className={cn('flex flex-col gap-6')}>
+            <ApiKeyForm
+              initialApiKeyData={initialApiKeyData}
+              initialAvailableScope={initialAvailableScope}
+              userRole={userRole}
+            />
+            <ApiKeyDisplay initialApiKeyData={initialApiKeyData} />
+          </div>
         </div>
       </main>
     </div>
