+++ apps/client/src/widgets/home/index.ts
@@ -1,5 +1,6 @@
 export { default as ApiKeyHeader } from './ui/ApiKeyHeader';
-export { default as ApiKeyCard } from './ui/ApiKeyCard';
+export { default as ApiKeyForm } from './ui/ApiKeyForm';
+export { default as ApiKeyDisplay } from './ui/ApiKeyDisplay';
 export * from './model/useCreateApiKey';
 export * from './model/useDeleteApiKey';
 export * from './model/useGetApiKey';
